requirements.txt
pytube
ffmpeg-python
argparse
pyqrcode